Move pat_suggest.go into `cli` package
