+++ cli/mobycli/pat_suggest_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package utils
+package mobycli
 
 import (
 	"testing"
